Make the install script badass

+++ setup/install.sh
@@ -12,41 +12,61 @@ echo "========================================================"
 echo "continue? Y/n"
 read answer
 
-echo $answer
-
 if [[ $answer == 'y'* ]]; then
     cd ~
     rm -rf ~/kreeper/
 
-    # TO DO: how to 
     git clone git@github.com:avacordero90/kreeper.git
 
-    cd ~/kreeper/
-    curl -s "https://www.python.org/ftp/python/3.9.6/Python-3.9.6.tgz" --output python.tgz && tar -zxvf python.tgz
-    sudo rm -rf python.tgz
+    # test that we have git/perms
+    if [[ ! $1 ]]; then
+        cd ~/kreeper/
 
-    cd ~/kreeper/Python-3.9.6
-    sudo make clean && sudo /configure --prefix=${HOME}/localpython --enable-optimizations
-    sudo make && sudo make install
+        curl -s "https://www.python.org/ftp/python/3.9.6/Python-3.9.6.tgz" -o python.tgz && tar -zxvf python.tgz
 
-    curl -s https://bootstrap.pypa.io/get-pip.py -o get-pip.py
-    pip --python /usr/bin/python3
+        if [[ ! $1 ]]; then
+            sudo rm -rf python.tgz
 
-    sudo python get-pip.py
-    rm -f ~/kreeper/get-pip.py
+            cd ~/kreeper/Python-3.9.6
+            sudo make clean
 
-    sudo -H pip install -U pipenv
-    # pipenv --python /bin/python3.8
+            sudo ./configure --prefix=${HOME}/localpython --enable-optimizations
+            
+            if [[ ! $1 ]]; then
+                
+                if [[ $(sudo make && sudo make install && \
+                curl -s https://bootstrap.pypa.io/get-pip.py -o get-pip.py && \
+                sudo python get-pip.py) ]]; then
+                
+                    rm -f ~/kreeper/get-pip.py
 
-    cd ~/kreeper/
-    pipenv clean
-    
-    pipenv shell source ./install/config.sh
+                    if [[ $(sudo -H pip install -U pipenv) ]]; then
+                        # pipenv --python /bin/python3.9
 
-    if [[ $(pipenv --version) ]]; then
-        echo -e "installation complete!\n"
+                        cd ~/kreeper/
+                        if [[ $(pipenv clean && pipenv shell source ./install/config.sh) ]]; then
 
-        ~/kreeper/setup/config.sh
+                            if [[ $(pipenv --version) ]]; then
+                                echo -e "installation complete!\n"
+                            else
+                            # TO DO: this goes somewhere else
+                                echo -e "installation failed: pipenv not found.\n"
+                            fi
+                        else
+                            echo -e "installation failed!\n"
+                        fi
+                    else
+                        echo -e "installation failed!\n"
+                    fi
+                else
+                    echo -e "installation failed!\n"
+                fi
+            else
+                echo -e "installation failed!\n"
+            fi
+        else
+            echo -e "installation failed!\n"
+        fi
     else
         echo -e "installation failed!\n"
     fi
